+++ domains/certificates/Query.API/API/ContractService/ICertificateGenerationSignUpService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace API.CertificateGenerationSignUpService;
+namespace API.ContractService;
 
 public interface ICertificateGenerationSignUpService
 {
