+++ domains/certificates/Query.API/API/ContractService/Repositories/ICertificateGenerationSignUpRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace API.CertificateGenerationSignUpService.Repositories;
+namespace API.ContractService.Repositories;
 
 public interface ICertificateGenerationSignUpRepository
 {
