+++ domains/certificates/Query.API/API/Query.API/Controllers/SignUpsController.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
-using API.CertificateGenerationSignUpService;
+using API.ContractService;
 using API.Query.API.ApiModels.Requests;
 using API.Query.API.ApiModels.Responses;
 using FluentValidation;
@@ -11,7 +11,7 @@ using FluentValidation.AspNetCore;
 using Marten;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using static API.CertificateGenerationSignUpService.CreateSignUpResult;
+using static API.ContractService.CreateSignUpResult;
 
 namespace API.Query.API.Controllers;
